[open-zaak/open-notificaties#231] Add catalogus kenmerk to besluiten & documenten kanaal.
