+++ src/openzaak/components/documenten/api/kanalen.py
@@ -1,6 +1,7 @@
 # SPDX-License-Identifier: EUPL-1.2
 # Copyright (C) 2019 - 2020 Dimpact
-from notifications_api_common.kanalen import Kanaal
+from openzaak.notifications.kanaal import Kanaal
+from openzaak.utils.help_text import mark_experimental
 
 from ..models import EnkelvoudigInformatieObject
 
@@ -11,5 +12,13 @@ KANAAL_DOCUMENTEN = Kanaal(
         "bronorganisatie",
         "informatieobjecttype",
         "vertrouwelijkheidaanduiding",
+        "informatieobjecttype.catalogus",
     ),
+    extra_kwargs={
+        "informatieobjecttype.catalogus": {
+            "help_text": mark_experimental(
+                "URL-referentie naar de CATALOGUS waartoe dit INFORMATIEOBJECTTYPE behoort."
+            )
+        }
+    },
 )
